2024/app/requirements.txt
glob2==0.7
Pillow==10.4.0
plotly==5.24.1
Requests==2.32.3
streamlit==1.39.0
